+++ src/common/display.ts
@@ -42,7 +42,7 @@ interface ScrollState {
 }
 
 export enum ScrollType { vertical = 0, horizontal = 1 }
-export enum UpdateType { none = 0, view = 1, overlays = 2, selection = 4, scrollbars = 8, update = 16, scroll = 32, scrollEnd = 64, scrollView = 128, display = 256, selectionChanged = 512, scrollViewOverlays = 1024, layout = 2048 }
+export enum UpdateType { none = 0, view = 1, overlays = 2, selection = 4, scrollbars = 8, update = 16, scroll = 32, scrollEnd = 64, scrollView = 128, display = 256, selectionChanged = 512, scrollViewOverlays = 1024, layout = 2048, scrollReset = 4096 }
 
 enum CornerType {
     Flat = 0,
@@ -129,7 +129,18 @@ export class Display extends EventEmitter {
 
     public split = null;
     public splitLive: boolean = false;
-    public scrollLock: boolean = false;
+    private _scrollLock: boolean = false;
+
+    get scrollLock(): boolean {
+        return this._scrollLock;
+    }
+    set scrollLock(locked: boolean) {
+        if (locked !== this._scrollLock) {
+            this._scrollLock = locked;
+            this._VScroll.autoScroll = !locked;
+        }
+    }
+
 
     private _linkFunction;
     private _mxpLinkFunction;
@@ -914,10 +925,14 @@ export class Display extends EventEmitter {
                     else
                         this.split.updatePosition();
                 }
+                this._updating |= UpdateType.scrollReset;
+                this._updating &= ~UpdateType.scroll;
+            }
+            if ((this._updating & UpdateType.scrollReset) === UpdateType.scrollReset) {
                 this._view.style.transform = `translate(${-this._HScroll.position}px, ${-this._VScroll.position}px)`;
                 this._background.style.transform = `translate(${-this._HScroll.position}px, ${-this._VScroll.position}px)`;
                 this._overlay.style.transform = `translate(${-this._HScroll.position}px, ${-this._VScroll.position}px)`;
-                this._updating &= ~UpdateType.scroll;
+                this._updating &= ~UpdateType.scrollReset;
             }
             if ((this._updating & UpdateType.view) === UpdateType.view) {
                 this.updateView();
@@ -1095,15 +1110,6 @@ export class Display extends EventEmitter {
             selection: []
         };
         this._viewCache = {};
-        if (this.split) {
-            this.split.viewCache = {};
-            this.split.dirty = true;
-            if (this.split.shown) {
-                this.split.style.display = '';
-                this.split.shown = false;
-                if (this._scrollCorner) this._scrollCorner.classList.remove('active');
-            }
-        }
         this.lineIDs = [];
         this._lines = [];
         this._lineID = 0;
@@ -1121,7 +1127,17 @@ export class Display extends EventEmitter {
         this._parser.Clear();
         this._VScroll.reset();
         this._HScroll.reset();
-        this.doUpdate(UpdateType.scrollbars);
+        if (this.split) {
+            this.split.viewCache = {};
+            this.split.dirty = true;
+            if (this.split.shown) {
+                this.split.style.display = '';
+                this.split.shown = false;
+                if (this._scrollCorner) this._scrollCorner.classList.remove('active');
+            }
+        }
+        this._updating &= ~UpdateType.scroll;
+        this.doUpdate(UpdateType.scrollbars | UpdateType.scrollReset);
     }
 
     public updateFont(font?: string, size?: string) {
@@ -1471,8 +1487,8 @@ export class Display extends EventEmitter {
         if (start >= lineLength) return;
         if (!start || start < 0) start = 0;
         if (!end || end > lineLength)
-            end = lineLength;        
-       if(start === end)
+            end = lineLength;
+        if (start === end)
             return;
         //clear cache
         if (this._viewCache[idx])
@@ -3696,6 +3712,10 @@ export class Display extends EventEmitter {
         this._VScroll.scrollToEnd();
     }
 
+    public get scrollAtBottom() {
+        return this._VScroll.atBottom;
+    }
+
     private expireLineLinkFormat(formats, idx: number) {
         let f;
         let fs;
@@ -4103,7 +4123,7 @@ export class ScrollBar extends EventEmitter {
             position: 0
         };
         this.maxPosition = 0;
-        this.updatePosition(0);
+        this.updatePosition(0, true);
         this.update();
     }
 
@@ -4268,19 +4288,21 @@ export class ScrollBar extends EventEmitter {
      * @memberof ScrollBar
      * @fires ScrollBar#scroll
      */
-    private updatePosition(p) {
+    private updatePosition(p, force?) {
         if (p < 0 || this._maxDrag < 0)
             p = 0;
         else if (p > this._maxDrag)
             p = this._maxDrag;
-
+        const prv = this.position;
         this.thumb.style[this._type === ScrollType.horizontal ? 'left' : 'top'] = p + 'px';
         this.state.dragPosition = p;
         this._position = p * this._ratio;
         if (this._position < 0)
             this._position = 0;
         this.update();
-        this.emit('scroll', this.position);
+        //only update if it changed
+        if (prv !== this.position || force)
+            this.emit('scroll', this.position);
     }
 
     /**